Prevent self damage
A GameObject may have several attack appendages attached to it. If the
body hits one of these, then we want to ignore that damage.

We could use a similar method to transfer damage to the parent:

    v = victim.gameObject.GetComponentInParent<Vitality>();

However, I don't think this is a good idea. You need to make direct
contact with something that has Vitality to damage it. If we want an
appendage to take damage, it should have Vitality. We can do custom
behavior in the OnTakeDamage call (however, I don't have a good way to
script that yet).