Rename frontCheck to attackVolume
I guess we use it for more than attacks, but I want it to be clear when
something has an Attack on it.

"Shape" may be more accurate than "Volume", but volume isn't confusing
since everything is 2D.

Any other changes are automatic from editing the prefabs. I didn't have
to add them to the scene to see their hierarchy -- it's visible if you
expand them in project view.