Support delegates in Vitality
Allow users to register their handlers instead of subclassing Vitality.
This makes it easier to add Vitality to an object and hook into
object-specific behavior with a bit less code. Reduces the amount of
xHealth classes like RocketHealth that are just wrappers around a
function. Removes RocketHealth.

You can pass null to either delegate argument and we'll fall back on the
empty implementations.

I'm using MixedCase for delegates so they look like functions.