Refactor Enemy to use Mob
This change makes the code a bit more complex, but has better re-use.

It would be cleaner to provide a movement controller component for each
type of Mob (player, clone, enemy) that the mob queries to determine how
to move. Or make that a separate level of subclass (instead of mingling
player details with the movement).

Enemies currently don't jump or do speed-based movement animations, so
we ignore jump if there's no groundCheck and add an unused Speed
AnimParam, respectively.

This also changes enemies to use force-based movement instead of just
setting their velocity.